Moved the tests in their own module
