+++ openquake/engine/tests/engine_test.py
@@ -21,15 +21,11 @@ import subprocess
 import tempfile
 import unittest
 import warnings
-import time
-import mock
-import signal
 
 from openquake.engine.db import models
 from django.core import exceptions
 
 from openquake.engine import engine
-from openquake.engine.celery_node_monitor import CeleryNodeMonitor
 from openquake.engine.tests.utils import helpers
 
 
@@ -437,53 +433,3 @@ class DeleteRiskCalcTestCase(unittest.TestCase):
         risk_calc = risk_job.risk_calculation
 
         self.assertRaises(RuntimeError, engine.del_risk_calc, risk_calc.id)
-
-
-class CeleryNodeMonitorTestCase(unittest.TestCase):
-    def setUp(self):
-        self.patch = mock.patch('celery.task.control.inspect')
-        self.inspect = self.patch.start()
-
-    def test_all_nodes_were_down(self):
-        ping = self.inspect().ping
-        ping.return_value = {}
-        mon = CeleryNodeMonitor(no_distribute=False, interval=0.1)
-        with self.assertRaises(SystemExit), mock.patch('sys.stderr') as stderr:
-            mon.__enter__()
-        self.assertEqual(ping.call_count, 1)  # called only once
-        self.assertTrue(stderr.write.called)  # an error message was printed
-
-    def test_all_nodes_are_up(self):
-        ping = self.inspect().ping
-        ping.return_value = {'node1': []}
-        mon = CeleryNodeMonitor(no_distribute=False, interval=0.1)
-        with mon:
-            time.sleep(.21)
-        # three ping were done in the thread, plus 1 at the beginning
-        self.assertEqual(ping.call_count, 4)
-
-    def test_one_node_went_down(self):
-        ping = self.inspect().ping
-        ping.return_value = {'node1': []}
-        mon = CeleryNodeMonitor(no_distribute=False, interval=0.1)
-        with mon, mock.patch('os.kill') as kill, \
-                mock.patch('sys.stderr') as stderr:
-            time.sleep(.11)
-            ping.return_value = {}
-            time.sleep(.1)
-            # two ping were done in the thread, plus 1 at the beginning
-            self.assertEqual(ping.call_count, 3)
-
-            # check that kill was called with a SIGABRT
-            pid, signum = kill.call_args[0]
-            self.assertEqual(pid, os.getpid())
-            self.assertEqual(signum, signal.SIGABRT)
-            self.assertTrue(stderr.write.called)
-
-    def test_no_distribute(self):
-        with CeleryNodeMonitor(no_distribute=True, interval=0.1):
-            time.sleep(0.5)
-        self.assertIsNone(self.inspect.call_args)
-
-    def tearDown(self):
-        self.patch.stop()
